Add feature `alloc`
`std` was the provider for implementations that depends on `Vec`.
However, that was an implementation shortcut, and the proper way to
handle that is to use `alloc` that is `no-std` compatible.

With `alloc`, the features/test structure is simplified. With this
change, all tests are no-std compatible, and can optionally run under
`std` too.