Add `width` check to `fmt::{Lower,Upper}Hex`
Its beneficial to have some truncation strategy for bytes representation
so they can be readable.

However, truncation alone may result in collisions and misleading
information.

This PR introduces the usage of `core::fmt::Formatter::width`, a Rust
standard for truncating formatted objects. It will reduce arbitrary
slices into smaller hex representations. These smaller representations
will be the result of the sum of XOR over the chunks of the original
bytes.

This is secure and will lead to unique representations, truncated only
to the available bits of the desired width.